requirements.txt
langchain==0.3.25
langchain_community==0.3.25
openai==1.86.0
python-dotenv==1.1.0
streamlit==1.45.1
toml==0.10.2
ruff==0.11.13
sentence-transformers==4.1.0
watchdog==6.0.0
pypdf==5.6.0
faiss-cpu==1.11.0
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
langchain-huggingface==0.3.0
